Fixes for vox-format integration:
 - Implement conversion traits for vox-format point type
 - Add `VoxChannel` trait to avoid orphan rules, and allow easier
   loading of custom channel types.
 - Test that loads a file into Array3x1<ColorIndex>
 - Docs
