+++ crates/building_blocks_core/Cargo.toml
@@ -27,3 +27,4 @@ glam = { version = "0.13.0", optional = true }
 mint = { version = "0.5.0", optional = true }
 nalgebra = { version = "0.27", optional = true }
 sdfu = { version = "0.3", optional = true }
+vox-format = { version = "0.1", optional = true }
