+++ crates/building_blocks_core/src/point.rs
@@ -7,6 +7,8 @@ mod glam_conversions;
 mod mint_conversions;
 #[cfg(feature = "nalgebra")]
 mod nalgebra_conversions;
+#[cfg(feature = "vox-format")]
+mod vox_format_conversions;
 #[cfg(feature = "sdfu")]
 mod sdfu_integration;
