+++ crates/building_blocks_storage/src/lib.rs
@@ -118,9 +118,7 @@ mod dot_vox_conversions;
 #[cfg(feature = "dot_vox")]
 pub use dot_vox_conversions::*;
 #[cfg(feature = "vox-format")]
-mod vox_format_conversions;
-#[cfg(feature = "vox-format")]
-pub use vox_format_conversions::*;
+pub mod vox_format;
 #[cfg(feature = "image")]
 mod image_conversions;
 #[cfg(feature = "image")]
